Split out mockTestRequests object to new file
