+++ src/adapters/mock/adapter.ts
@@ -6,6 +6,8 @@ import {
   NormalizedResponse,
 } from '../../runtime/http';
 
+import {mockTestRequests} from './mock_test_requests';
+
 interface MockAdapterArgs extends AdapterArgs {
   rawRequest: NormalizedRequest;
 }
@@ -30,86 +32,30 @@ export async function mockFetch({
   body,
 }: NormalizedRequest): Promise<NormalizedResponse> {
   canonicalizeHeaders(headers);
-  const matchingRequest = mockAdapter.findRequest({
+  const matchingRequest = mockTestRequests.findRequest({
     url,
     method,
     headers,
     body,
   });
-  if (typeof matchingRequest === 'undefined') {
-    mockAdapter.requestList.push({url, method, headers, body, attempts: 1});
+  if (matchingRequest === undefined) {
+    mockTestRequests.requestList.push({
+      url,
+      method,
+      headers,
+      body,
+      attempts: 1,
+    });
   } else {
     matchingRequest.attempts++;
   }
 
-  const next = mockAdapter.responseList.shift()!;
+  const next = mockTestRequests.responseList.shift()!;
   if (next instanceof Error) {
     throw next;
   }
   return next;
 }
-
-type RequestListEntry = NormalizedRequest & {attempts: number};
-type ResponseListEntry = NormalizedResponse | Error;
-
-interface MockedAdapter {
-  requestList: RequestListEntry[];
-  responseList: ResponseListEntry[];
-  getOldestRequest: () => NormalizedRequest;
-  getMostRecentRequest: () => NormalizedRequest;
-  queueResponse: (response: NormalizedResponse) => void;
-  queueError: (error: Error) => void;
-  findRequest: (request: NormalizedRequest) => RequestListEntry | undefined;
-  reset: () => void;
-}
-
-export const mockAdapter: MockedAdapter = {
-  requestList: [],
-  responseList: [],
-
-  getOldestRequest(): NormalizedRequest {
-    if (this.requestList.length === 0) {
-      throw new Error('No requests have been made');
-    }
-    return this.requestList.shift()!;
-  },
-
-  getMostRecentRequest(): NormalizedRequest {
-    if (this.requestList.length === 0) {
-      throw new Error('No requests have been made');
-    }
-    return this.requestList.pop()!;
-  },
-
-  queueResponse(response: NormalizedResponse): void {
-    this.responseList.push(response);
-  },
-
-  queueError(error: Error): void {
-    this.responseList.push(error);
-  },
-
-  findRequest(request: NormalizedRequest): RequestListEntry | undefined {
-    for (const matchingRequest of this.requestList) {
-      if (
-        matchingRequest.url === request.url &&
-        matchingRequest.method === request.method
-      ) {
-        if (request.body === undefined || request.body === null) {
-          return matchingRequest;
-        } else if (matchingRequest.body === request.body) {
-          return matchingRequest;
-        }
-      }
-    }
-    return undefined;
-  },
-
-  reset() {
-    this.requestList = [];
-    this.responseList = [];
-  },
-};
 
 export function mockCreateDefaultStorage() {
   return new MemorySessionStorage();
